:recycle: Separate file for IndexStack
